refactor:upgrade InMemoryManagerTest
